api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0
redis==5.0.1
asyncpg==0.29.0
alembic==1.13.0
sqlalchemy==2.0.23